Update to latest primogenitor
This updates to the latest primogenitor release and makes lots of
small improvements suggested by SpotBugs. It also switches to using
jxe for increased security during XML parsing. It adds "published"
fields to the Atom feeds (which were created but never attached to
the tree, during serialization due to a bug).